chore(app): move tests to respective modules
