+++ reader/src/test/java/fr/carbonit/th/command/AdventurerCommandTest.java
