+++ reader/src/test/java/fr/carbonit/th/command/MapCommandTest.java
