+++ reader/src/test/java/fr/carbonit/th/command/TreasureCommandTest.java
