+++ reader/src/test/java/fr/carbonit/th/configuration/TreasureHuntConfigurationTest.java
