+++ reader/src/test/java/fr/carbonit/th/configuration/TreasureTest.java
