+++ reader/src/test/java/fr/carbonit/th/parser/ConfigurationParserTest.java
