+++ reader/src/test/java/fr/carbonit/th/provider/ConfigurationProvider.java
