+++ reader/src/test/java/fr/carbonit/th/reader/ConfigurationReaderTest.java
