+++ reader/src/test/java/fr/carbonit/th/reader/ReaderFileType.java
