+++ simulation/src/test/java/fr/carbonit/th/PositionTest.java
@@ -1,4 +1,4 @@
-package fr.carbonit.th.simulation;
+package fr.carbonit.th;
 
 import fr.carbonit.th.localisation.Coordinates;
 import fr.carbonit.th.localisation.Orientation;
